Fix review discussions

+++ cloud/disk_manager/internal/pkg/dataplane/snapshot/storage/storage_ydb_impl.go
@@ -376,7 +376,7 @@ func (s *storageYDB) snapshotCreated(
 		return err
 	}
 
-	state.encryptionMode = encryptionMode
+	state.encryptionMode = uint32(encryptionMode)
 	state.encryptionKeyHash = encryptionKeyHash
 
 	_, err = tx.Execute(ctx, fmt.Sprintf(`
