+++ cloud/disk_manager/internal/pkg/resources/images.go
@@ -427,7 +427,7 @@ func (s *storageYDB) createImage(
 		return ImageMeta{}, err
 	}
 
-	state.encryptionMode = encryptionMode
+	state.encryptionMode = uint32(encryptionMode)
 	state.encryptionKeyHash = encryptionKeyHash
 
 	_, err = tx.Execute(ctx, fmt.Sprintf(`
